make sure one can write in the area before trying to delete stuff
git-svn-id: svn+ssh://svn.cern.ch/reps/dirac/ILCDIRAC/trunk/ILCDIRAC@49185 4525493e-7705-40b1-a816-d608a930855b

+++ Core/Utilities/TARsoft.py
@@ -59,27 +59,27 @@ def install(app,config,area):
 
   if os.path.exists(folder_name):
     # and not appName =="slic":
+    appli_exists = True
     if not overwrite:
       DIRAC.gLogger.info("Folder or file %s found in %s, skipping install !"%(folder_name,area))
     else:
       DIRAC.gLogger.info("Overwriting %s found in %s"%(folder_name,area))
-      DIRAC.gLogger.info("First we delete existing version %s"%folder_name)
-      if os.path.exists(folder_name):
-        if os.path.isdir(folder_name):
-          try:
-            shutil.rmtree(folder_name)
-          except Exception,x:
-            DIRAC.gLogger.error("Failed deleting %s because %s,%s"%(folder_name,Exception,str(x)))
-        else:
-          try:
-            os.remove(folder_name)
-          except Exception,x:
-            DIRAC.gLogger.error("Failed deleting %s because %s,%s"%(folder_name,Exception,str(x)))
-      if os.path.exists(folder_name):
-        DIRAC.gLogger.error("Oh Oh, something was not right, the directory %s is still here"%folder_name)  
-    appli_exists = True
-    if overwrite:
       appli_exists = False
+      if CanWrite(area):
+        DIRAC.gLogger.info("First we delete existing version %s"%folder_name)
+        if os.path.exists(folder_name):
+          if os.path.isdir(folder_name):
+            try:
+              shutil.rmtree(folder_name)
+            except Exception,x:
+              DIRAC.gLogger.error("Failed deleting %s because %s,%s"%(folder_name,Exception,str(x)))
+          else:
+            try:
+              os.remove(folder_name)
+            except Exception,x:
+              DIRAC.gLogger.error("Failed deleting %s because %s,%s"%(folder_name,Exception,str(x)))
+        if os.path.exists(folder_name):
+          DIRAC.gLogger.error("Oh Oh, something was not right, the directory %s is still here"%folder_name)  
     #os.chdir(curdir)
     #return DIRAC.S_OK()
   if not appli_exists:
